tests: devolve OIDC model tests
