+++ tests/unit/oidc/models/test_base.py
@@ -1,0 +1,29 @@
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import pretend
+
+from warehouse.oidc.models import base
+
+
+def test_check_claim_binary():
+    wrapped = base._check_claim_binary(str.__eq__)
+
+    assert wrapped("foo", "bar", pretend.stub()) is False
+    assert wrapped("foo", "foo", pretend.stub()) is True
+
+
+class TestOIDCPublisher:
+    def test_oidc_publisher_not_default_verifiable(self):
+        publisher = base.OIDCPublisher(projects=[])
+
+        assert not publisher.verify_claims(signed_claims={})
